+++ iceoryx_posh/include/iceoryx_posh/internal/popo/untyped_publisher_impl.hpp
@@ -18,7 +18,7 @@
 #ifndef IOX_POSH_POPO_UNTYPED_PUBLISHER_IMPL_HPP
 #define IOX_POSH_POPO_UNTYPED_PUBLISHER_IMPL_HPP
 
-#include "iceoryx_posh/popo/base_publisher.hpp"
+#include "iceoryx_posh/internal/popo/base_publisher.hpp"
 #include "iceoryx_posh/popo/sample.hpp"
 
 namespace iox
